tcp traffic: refactoring
Refactoring of the TCP traffic sampler to combine the procfs
sampler and bpf sampler depending on enabled features.
