+++ src/samplers/tcp/linux/traffic/mod.rs
@@ -1,146 +1,40 @@
+use crate::common::Nop;
+use crate::samplers::tcp::*;
+
+const NAME: &str = "tcp_traffic";
+
+#[cfg(feature = "bpf")]
+mod bpf;
+
+mod proc;
+
+#[cfg(feature = "bpf")]
+use bpf::*;
+
+use proc::*;
+
+#[cfg(feature = "bpf")]
 #[distributed_slice(TCP_SAMPLERS)]
 fn init(config: &Config) -> Box<dyn Sampler> {
-    if let Ok(s) = Traffic::new(config) {
+    // try to initialize the bpf based sampler
+    if let Ok(s) = TcpTraffic::new(config) {
+        Box::new(s)
+    // try to fallback to the /proc/net/snmp based sampler if there was an error
+    } else if let Ok(s) = ProcNetSnmp::new(config) {
         Box::new(s)
     } else {
         Box::new(Nop {})
     }
 }
 
-mod bpf {
-    include!(concat!(env!("OUT_DIR"), "/tcp_traffic.bpf.rs"));
-}
-
-const NAME: &str = "tcp_traffic";
-
-use bpf::*;
-
-use crate::common::bpf::*;
-use crate::common::*;
-use crate::samplers::tcp::stats::*;
-use crate::samplers::tcp::*;
-
-impl GetMap for ModSkel<'_> {
-    fn map(&self, name: &str) -> &libbpf_rs::Map {
-        self.obj.map(name).unwrap()
-    }
-}
-
-/// Collects TCP Traffic stats using BPF and traces:
-/// * `tcp_sendmsg`
-/// * `tcp_cleanup_rbuf`
-///
-/// And produces these stats:
-/// * `tcp/receive/bytes`
-/// * `tcp/receive/segments`
-/// * `tcp/receive/size`
-/// * `tcp/transmit/bytes`
-/// * `tcp/transmit/segments`
-/// * `tcp/transmit/size`
-pub struct Traffic {
-    bpf: Bpf<ModSkel<'static>>,
-    counter_interval: Duration,
-    counter_next: Instant,
-    counter_prev: Instant,
-    distribution_interval: Duration,
-    distribution_next: Instant,
-    distribution_prev: Instant,
-}
-
-impl Traffic {
-    pub fn new(config: &Config) -> Result<Self, ()> {
-        // check if sampler should be enabled
-        if !config.enabled(NAME) {
-            return Err(());
-        }
-
-        let builder = ModSkelBuilder::default();
-        let mut skel = builder
-            .open()
-            .map_err(|e| error!("failed to open bpf builder: {e}"))?
-            .load()
-            .map_err(|e| error!("failed to load bpf program: {e}"))?;
-
-        skel.attach()
-            .map_err(|e| error!("failed to attach bpf program: {e}"))?;
-
-        let mut bpf = Bpf::from_skel(skel);
-
-        let counters = vec![
-            Counter::new(&TCP_RX_BYTES, Some(&TCP_RX_BYTES_HISTOGRAM)),
-            Counter::new(&TCP_TX_BYTES, Some(&TCP_TX_BYTES_HISTOGRAM)),
-            Counter::new(&TCP_RX_SEGMENTS, Some(&TCP_RX_SEGMENTS_HISTOGRAM)),
-            Counter::new(&TCP_TX_SEGMENTS, Some(&TCP_TX_SEGMENTS_HISTOGRAM)),
-        ];
-
-        bpf.add_counters("counters", counters);
-
-        let mut distributions = vec![("rx_size", &TCP_RX_SIZE), ("tx_size", &TCP_TX_SIZE)];
-
-        for (name, histogram) in distributions.drain(..) {
-            bpf.add_distribution(name, histogram);
-        }
-
-        Ok(Self {
-            bpf,
-            counter_interval: config.interval(NAME),
-            counter_next: Instant::now(),
-            counter_prev: Instant::now(),
-            distribution_interval: config.distribution_interval(NAME),
-            distribution_next: Instant::now(),
-            distribution_prev: Instant::now(),
-        })
-    }
-
-    pub fn refresh_counters(&mut self, now: Instant) {
-        if now < self.counter_next {
-            return;
-        }
-
-        let elapsed = (now - self.counter_prev).as_secs_f64();
-
-        self.bpf.refresh_counters(elapsed);
-
-        // determine when to sample next
-        let next = self.counter_next + self.counter_interval;
-
-        // check that next sample time is in the future
-        if next > now {
-            self.counter_next = next;
-        } else {
-            self.counter_next = now + self.counter_interval;
-        }
-
-        // mark when we last sampled
-        self.counter_prev = now;
-    }
-
-    pub fn refresh_distributions(&mut self, now: Instant) {
-        if now < self.distribution_next {
-            return;
-        }
-
-        self.bpf.refresh_distributions();
-
-        // determine when to sample next
-        let next = self.distribution_next + self.distribution_interval;
-
-        // check that next sample time is in the future
-        if next > now {
-            self.distribution_next = next;
-        } else {
-            self.distribution_next = now + self.distribution_interval;
-        }
-
-        // mark when we last sampled
-        self.distribution_prev = now;
-    }
-}
-
-impl Sampler for Traffic {
-    fn sample(&mut self) {
-        let now = Instant::now();
-        self.refresh_counters(now);
-        self.refresh_distributions(now);
+#[cfg(not(feature = "bpf"))]
+#[distributed_slice(CPU_SAMPLERS)]
+fn init(config: &Config) -> Box<dyn Sampler> {
+    // try to use the /proc/net/snmp based sampler since BPF was not enabled for
+    // this build
+    if let Ok(s) = ProcNetSnmp::new(config) {
+        Box::new(s)
+    } else {
+        Box::new(Nop {})
     }
 }
